feat(assembly): 添加指针类型和相关操作支持
- 在 PrimitiveType 枚举中添加了指针类型
- 在 GenCode 类中实现了与指针相关的操作，如取地址、解引用等
- 更新了 ExprNode 类及其子类，增加了对指针类型的支持
- 优化了类型计算和转换逻辑，以支持指针运算

+++ src/scanner/scanner.h
@@ -32,7 +32,7 @@ class Scanner {
 
         void skip();
         void putback(char c);
-        int scanint(int c);
+        long scanint(int c);
         char next();
         char peek();
         void scanNumeric(Token& token, char c);
